+++ packages/expo-brownfield/cli/src/commands/build-ios.ts
@@ -2,28 +2,27 @@ import type { Command } from 'commander';
 
 import {
   buildFramework,
-  cleanUpArtifacts,
-  createSwiftPackage,
-  createXcframework,
-  copyHermesXcframework,
-  makeArtifactsDirectory,
   printIosConfig,
   resolveBuildConfigIos,
   validatePrebuild,
+  shipSwiftPackage,
+  shipFrameworks,
 } from '../utils';
 
 const buildIos = async (command: Command) => {
   await validatePrebuild('ios');
-
   const config = resolveBuildConfigIos(command.opts());
   printIosConfig(config);
 
-  await cleanUpArtifacts(config);
-  makeArtifactsDirectory(config);
   await buildFramework(config);
-  await createSwiftPackage(config);
-  await createXcframework(config);
-  await copyHermesXcframework(config);
+
+  if (config.output !== 'frameworks') {
+    // Ship frameworks as swift package
+    shipSwiftPackage(config);
+  } else {
+    // Ship frameworks as standalone XCFrameworks
+    shipFrameworks(config);
+  }
 };
 
 export default buildIos;