+++ packages/expo-dev-launcher/ios/ReactDelegateHandler/ExpoDevLauncherReactDelegateHandler.swift
@@ -6,66 +6,25 @@ import React
 
 private class DevLauncherWrapperView: UIView {
   weak var devLauncherViewController: UIViewController?
-  private var devLauncherConstraints: [NSLayoutConstraint] = []
 
-  #if os(iOS)
-  @objc
-  func orientationDidChange() {
-    if let controller = devLauncherViewController {
-      setDevLauncherViewControllerConstraints(controller)
-    }
-  }
-  #endif
-
-  func setDevLauncherViewControllerConstraints(_ viewController: UIViewController) {
-    viewController.view.translatesAutoresizingMaskIntoConstraints = false
-    if !devLauncherConstraints.isEmpty {
-      NSLayoutConstraint.deactivate(devLauncherConstraints)
-    }
-    devLauncherConstraints = [
-      viewController.view.topAnchor.constraint(equalTo: self.topAnchor),
-      viewController.view.leadingAnchor.constraint(equalTo: self.leadingAnchor),
-      viewController.view.trailingAnchor.constraint(equalTo: self.trailingAnchor),
-      viewController.view.bottomAnchor.constraint(equalTo: self.bottomAnchor)
-    ]
-    NSLayoutConstraint.activate(devLauncherConstraints)
+  func setupDevLauncherView(_ viewController: UIViewController) {
+#if os(macOS)
+    viewController.view.autoresizingMask = [.width, .height]
+#else
+    viewController.view.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+#endif
+    viewController.view.frame = bounds
   }
 
 #if !os(macOS)
-  override func didMoveToWindow() {
-    super.didMoveToWindow()
-    #if os(iOS)
-    if window != nil {
-      NotificationCenter.default.addObserver(self, selector: #selector(orientationDidChange), name: UIDevice.orientationDidChangeNotification, object: nil)
-    }
-    #endif
-
-    guard let devLauncherViewController,
-      let window,
-      let rootViewController = window.rootViewController else {
-      return
-    }
-
-    let isSwiftUIController = NSStringFromClass(type(of: rootViewController)).contains("UIHostingController")
-    // TODO(pmleczek): Revisit this for a more reliable solution
-    let isBrownfield = NSStringFromClass(type(of: rootViewController)).contains("UINavigationController")
-    if !isSwiftUIController && !isBrownfield && devLauncherViewController.parent != rootViewController {
-      rootViewController.addChild(devLauncherViewController)
-      devLauncherViewController.didMove(toParent: rootViewController)
-      devLauncherViewController.view.setNeedsLayout()
-      devLauncherViewController.view.layoutIfNeeded()
-    }
+  override func layoutSubviews() {
+    super.layoutSubviews()
+    devLauncherViewController?.view.frame = bounds
   }
 
-  override func willMove(toWindow newWindow: UIWindow?) {
-    super.willMove(toWindow: newWindow)
-    if newWindow == nil {
-      devLauncherViewController?.willMove(toParent: nil)
-      devLauncherViewController?.removeFromParent()
-      #if os(iOS)
-      NotificationCenter.default.removeObserver(self, name: UIDevice.orientationDidChangeNotification, object: nil)
-      #endif
-    }
+  override func safeAreaInsetsDidChange() {
+    super.safeAreaInsetsDidChange()
+    devLauncherViewController?.view.layoutIfNeeded()
   }
 #endif
 }
@@ -110,7 +69,7 @@ public class ExpoDevLauncherReactDelegateHandler: ExpoReactDelegateHandler, EXDe
     let wrapperView = DevLauncherWrapperView()
     wrapperView.devLauncherViewController = viewController
     wrapperView.addSubview(viewController.view)
-    wrapperView.setDevLauncherViewControllerConstraints(viewController)
+    wrapperView.setupDevLauncherView(viewController)
     return wrapperView
   }