+++ packages/expo-symbols/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- Add explicit type re-export and return type annotations ([#43562](https://github.com/expo/expo/pull/43562) by [@kitten](https://github.com/kitten))
+
 ## 55.0.4 — 2026-02-16
 
 _This version does not introduce any user-facing changes._